+++ src/frontend/apps/drive/src/features/explorer/hooks/useMutations.ts
@@ -39,6 +39,24 @@ export const useMutationCreateFile = () => {
   });
 };
 
+export const useMutationCreateFileFromTemplate = () => {
+  const driver = getDriver();
+  const refresh = useRefreshQueryCacheAfterMutation();
+  return useMutation({
+    mutationFn: async (
+      ...payload: Parameters<typeof driver.createFileFromTemplate>
+    ) => {
+      return driver.createFileFromTemplate(...payload);
+    },
+    onSuccess: (data, variables) => {
+      refresh(variables.parentId);
+    },
+    meta: {
+      showErrorOn403: true,
+    },
+  });
+};
+
 export const useMutationDeleteItems = () => {
   const driver = getDriver();
   const { item } = useGlobalExplorer();